alembic/versions/c37c16f08ab4_init_migration.py
"""Init migration

Revision ID: c37c16f08ab4
Revises: 
Create Date: 2023-11-26 01:52:18.247067

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = 'c37c16f08ab4'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    pass


def downgrade() -> None:
    pass
