requirements_dev.txt
-r requirements.txt

blinker==1.6.3
click==8.1.7
pytest==7.4.3
Faker==18.13.0
factory-boy==3.3.0
pytest-cov==4.1.0
psycopg2-binary==2.9.9


